+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [5.5.0] - 2024-07-31
 
+### Added
+- Support for application command context types
+
 ### Changed
 - Upgraded Kotlin to 2.0.0
 - Upgraded to kotlinx.serialization 1.7.0